Version 5.5.2

+++ ireport-designer/src/com/jaspersoft/ireport/designer/palette/actions/DefaultCellElementsLayout.java
@@ -160,7 +160,7 @@ public class DefaultCellElementsLayout {
                     {
                         JRStyledTextParser styledTextParser = JRStyledTextParser.getInstance();
                         JRDesignTextElement dte = (JRDesignTextElement)element;
-                        dte.setFontSize(null);
+                        dte.setFontSize((Integer)null);
                         for (int i=dte.getFontSize()-1; i>1 ; --i)
                         {
                                 String text = "test";
